Remove obsolete test files and refactor existing tests for improved organization
- Deleted outdated test files including server-coverage.test.ts, server-client.test.ts, auth.test.ts, and server-auth.test.ts to streamline the codebase.
- Refactored existing tests to enhance clarity and maintainability, ensuring they follow updated patterns and practices.
- Improved error handling in authentication tests, ensuring comprehensive coverage for various authentication scenarios.
- Updated mock implementations to accurately simulate Azure DevOps API interactions, enhancing test reliability.
