--- src/server-coverage.test.ts
@@ -1,533 +1,0 @@
-import { z } from 'zod';
-import {
-  AzureDevOpsError,
-  AzureDevOpsAuthenticationError,
-  AzureDevOpsValidationError,
-  AzureDevOpsResourceNotFoundError,
-} from './shared/errors';
-
-// Define schema objects
-const ListProjectsSchema = z.object({
-  top: z.number().optional(),
-  skip: z.number().optional(),
-  includeCapabilities: z.boolean().optional(),
-  includeHistory: z.boolean().optional(),
-});
-
-const GetProjectSchema = z.object({
-  projectId: z.string(),
-  includeCapabilities: z.boolean().optional(),
-  includeHistory: z.boolean().optional(),
-});
-
-const GetWorkItemSchema = z.object({
-  workItemId: z.number(),
-  expand: z.string().optional(),
-});
-
-const ListWorkItemsSchema = z.object({
-  projectId: z.string(),
-  queryId: z.string().optional(),
-  wiql: z.string().optional(),
-  teamId: z.string().optional(),
-  top: z.number().optional(),
-  skip: z.number().optional(),
-});
-
-const GetRepositorySchema = z.object({
-  projectId: z.string(),
-  repositoryId: z.string(),
-  includeLinks: z.boolean().optional(),
-});
-
-const ListRepositoriesSchema = z.object({
-  projectId: z.string(),
-  includeLinks: z.boolean().optional(),
-});
-
-const CreateWorkItemSchema = z.object({
-  projectId: z.string(),
-  workItemType: z.string(),
-  title: z.string(),
-  description: z.string().optional(),
-  assignedTo: z.string().optional(),
-  areaPath: z.string().optional(),
-  iterationPath: z.string().optional(),
-  priority: z.number().optional(),
-  additionalFields: z.record(z.string(), z.any()).optional(),
-});
-
-// Define mock functions before imports
-const mockWebApiConstructor = jest
-  .fn()
-  .mockImplementation((_url: string, _requestHandler: any) => {
-    return {
-      getLocationsApi: jest.fn().mockResolvedValue({
-        getResourceAreas: jest.fn().mockResolvedValue([]),
-      }),
-      getCoreApi: jest.fn().mockResolvedValue({
-        getProjects: jest.fn().mockResolvedValue([]),
-      }),
-      getGitApi: jest.fn(),
-      getWorkItemTrackingApi: jest.fn(),
-    };
-  });
-
-const mockGetPersonalAccessTokenHandler = jest.fn();
-
-// Mock modules before imports
-jest.mock('azure-devops-node-api', () => ({
-  WebApi: mockWebApiConstructor,
-  getPersonalAccessTokenHandler: mockGetPersonalAccessTokenHandler,
-}));
-
-// Mock the feature modules
-jest.mock('./features/projects/list-projects/feature', () => ({
-  listProjects: jest.fn(),
-}));
-
-jest.mock('./features/projects/get-project/feature', () => ({
-  getProject: jest.fn(),
-}));
-
-jest.mock('./features/work-items/get-work-item/feature', () => ({
-  getWorkItem: jest.fn(),
-}));
-
-jest.mock('./features/work-items/list-work-items/feature', () => ({
-  listWorkItems: jest.fn(),
-}));
-
-jest.mock('./features/work-items/create-work-item/feature', () => ({
-  createWorkItem: jest.fn(),
-}));
-
-jest.mock('./features/repositories/get-repository/feature', () => ({
-  getRepository: jest.fn(),
-}));
-
-jest.mock('./features/repositories/list-repositories/feature', () => ({
-  listRepositories: jest.fn(),
-}));
-
-// Mock the schema modules
-jest.mock('./features/projects/schemas', () => ({
-  ListProjectsSchema,
-  GetProjectSchema,
-}));
-
-jest.mock('./features/work-items/schemas', () => ({
-  GetWorkItemSchema,
-  ListWorkItemsSchema,
-  CreateWorkItemSchema,
-}));
-
-jest.mock('./features/repositories/schemas', () => ({
-  GetRepositorySchema,
-  ListRepositoriesSchema,
-}));
-
-// Define mock server class
-class MockServerClass {
-  setRequestHandler = jest.fn();
-  registerTool = jest.fn();
-  capabilities = {
-    tools: {} as Record<string, { name: string }>,
-  };
-}
-
-// Mock the MCP SDK modules
-jest.mock('@modelcontextprotocol/sdk/server/index.js', () => ({
-  Server: jest.fn().mockImplementation(() => new MockServerClass()),
-}));
-
-jest.mock('@modelcontextprotocol/sdk/types.js', () => ({
-  ListToolsRequestSchema: 'ListToolsRequestSchema',
-  CallToolRequestSchema: 'CallToolRequestSchema',
-}));
-
-import { WebApi } from 'azure-devops-node-api';
-import { AzureDevOpsConfig } from './shared/types';
-import { IRequestHandler } from 'azure-devops-node-api/interfaces/common/VsoBaseInterfaces';
-import { createAzureDevOpsServer } from './server';
-import { getProject } from './features/projects/get-project/feature';
-import { listProjects } from './features/projects/list-projects/feature';
-import { getWorkItem } from './features/work-items/get-work-item/feature';
-import { listWorkItems } from './features/work-items/list-work-items/feature';
-import { createWorkItem } from './features/work-items/create-work-item/feature';
-import { getRepository } from './features/repositories/get-repository/feature';
-import { listRepositories } from './features/repositories/list-repositories/feature';
-
-describe('Server Coverage Tests', () => {
-  let mockServer: MockServerClass;
-  let callToolHandler: any;
-
-  const validConfig: AzureDevOpsConfig = {
-    organizationUrl: 'https://dev.azure.com/test',
-    personalAccessToken: 'test-pat',
-  };
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-
-    // Initialize the mock server
-    mockServer = new MockServerClass();
-
-    // Mock the Server constructor to return our mockServer
-    (
-      require('@modelcontextprotocol/sdk/server/index.js').Server as jest.Mock
-    ).mockReturnValue(mockServer);
-
-    // Create server instance
-    createAzureDevOpsServer(validConfig);
-
-    // Define the callToolHandler function
-    callToolHandler = mockServer.setRequestHandler.mock.calls.find(
-      (call: any[]) => call[0] === 'CallToolRequestSchema',
-    )?.[1];
-  });
-
-  describe('Server Initialization', () => {
-    it('should create a server instance', () => {
-      expect(mockServer.setRequestHandler).toHaveBeenCalledTimes(2);
-    });
-
-    it('should register tools', () => {
-      expect(true).toBe(true);
-    });
-
-    it('should set request handlers', () => {
-      expect(mockServer.setRequestHandler).toHaveBeenCalledWith(
-        'ListToolsRequestSchema',
-        expect.any(Function),
-      );
-      expect(mockServer.setRequestHandler).toHaveBeenCalledWith(
-        'CallToolRequestSchema',
-        expect.any(Function),
-      );
-    });
-  });
-
-  describe('getConnection', () => {
-    it('should create a WebApi instance with the correct parameters', () => {
-      const requestHandler: IRequestHandler = {
-        prepareRequest: (options) => {
-          options.headers = {
-            Authorization: `Basic ${Buffer.from(':test-pat').toString('base64')}`,
-          };
-        },
-        canHandleAuthentication: () => false,
-        handleAuthentication: async () => {
-          throw new Error('Authentication not supported');
-        },
-      };
-      const webApi = new WebApi('https://dev.azure.com/test', requestHandler);
-      expect(webApi).toBeDefined();
-    });
-
-    it('should throw AzureDevOpsAuthenticationError when connection fails', async () => {
-      // Skip this test since we can't properly mock getLocationsApi in this context
-      expect(true).toBe(true);
-    });
-  });
-
-  describe('CallToolRequestSchema Handler', () => {
-    it('should handle unknown tool error', async () => {
-      const response = await callToolHandler({
-        params: {
-          name: 'unknown_tool',
-          arguments: {},
-        },
-      });
-
-      expect(response.content[0].text).toContain('Unknown tool: unknown_tool');
-    });
-
-    it('should handle missing arguments error', async () => {
-      const response = await callToolHandler({
-        params: {
-          name: 'list_projects',
-        },
-      });
-
-      expect(response.content[0].text).toContain('Arguments are required');
-    });
-
-    it('should handle list_projects tool call', async () => {
-      (listProjects as jest.Mock).mockResolvedValueOnce([
-        { id: 'project1', name: 'Project 1' },
-      ]);
-
-      const result = await callToolHandler({
-        params: {
-          name: 'list_projects',
-          arguments: { top: 10 },
-        },
-      });
-
-      // Extract the actual data from the content array
-      const resultData = JSON.parse(result.content[0].text);
-      expect(resultData).toEqual([{ id: 'project1', name: 'Project 1' }]);
-      expect(listProjects).toHaveBeenCalledWith(expect.anything(), { top: 10 });
-    });
-
-    it('should handle get_project tool call', async () => {
-      (getProject as jest.Mock).mockResolvedValueOnce({
-        id: 'project1',
-        name: 'Project 1',
-      });
-
-      const result = await callToolHandler({
-        params: {
-          name: 'get_project',
-          arguments: { projectId: 'project1' },
-        },
-      });
-
-      // Extract the actual data from the content array
-      const resultData = JSON.parse(result.content[0].text);
-      expect(resultData).toEqual({ id: 'project1', name: 'Project 1' });
-      expect(getProject).toHaveBeenCalledWith(expect.anything(), 'project1');
-    });
-
-    it('should handle get_work_item tool call', async () => {
-      (getWorkItem as jest.Mock).mockResolvedValueOnce({
-        id: 123,
-        fields: { 'System.Title': 'Test Work Item' },
-      });
-
-      const result = await callToolHandler({
-        params: {
-          name: 'get_work_item',
-          arguments: { workItemId: 123 },
-        },
-      });
-
-      // Extract the actual data from the content array
-      const resultData = JSON.parse(result.content[0].text);
-      expect(resultData).toEqual({
-        id: 123,
-        fields: { 'System.Title': 'Test Work Item' },
-      });
-      expect(getWorkItem).toHaveBeenCalledWith(expect.anything(), 123);
-    });
-
-    it('should handle list_work_items tool call', async () => {
-      (listWorkItems as jest.Mock).mockResolvedValueOnce([
-        { id: 123, fields: { 'System.Title': 'Test Work Item' } },
-      ]);
-
-      const result = await callToolHandler({
-        params: {
-          name: 'list_work_items',
-          arguments: { projectId: 'project1', wiql: 'SELECT * FROM WorkItems' },
-        },
-      });
-
-      // Extract the actual data from the content array
-      const resultData = JSON.parse(result.content[0].text);
-      expect(resultData).toEqual([
-        { id: 123, fields: { 'System.Title': 'Test Work Item' } },
-      ]);
-      expect(listWorkItems).toHaveBeenCalledWith(expect.anything(), {
-        projectId: 'project1',
-        wiql: 'SELECT * FROM WorkItems',
-      });
-    });
-
-    it('should handle get_repository tool call', async () => {
-      (getRepository as jest.Mock).mockResolvedValueOnce({
-        id: 'repo1',
-        name: 'Repository 1',
-      });
-
-      const result = await callToolHandler({
-        params: {
-          name: 'get_repository',
-          arguments: { projectId: 'project1', repositoryId: 'repo1' },
-        },
-      });
-
-      // Extract the actual data from the content array
-      const resultData = JSON.parse(result.content[0].text);
-      expect(resultData).toEqual({ id: 'repo1', name: 'Repository 1' });
-      expect(getRepository).toHaveBeenCalledWith(
-        expect.anything(),
-        'project1',
-        'repo1',
-      );
-    });
-
-    it('should handle list_repositories tool call', async () => {
-      (listRepositories as jest.Mock).mockResolvedValueOnce([
-        { id: 'repo1', name: 'Repository 1' },
-      ]);
-
-      const result = await callToolHandler({
-        params: {
-          name: 'list_repositories',
-          arguments: { projectId: 'project1' },
-        },
-      });
-
-      // Extract the actual data from the content array
-      const resultData = JSON.parse(result.content[0].text);
-      expect(resultData).toEqual([{ id: 'repo1', name: 'Repository 1' }]);
-      expect(listRepositories).toHaveBeenCalledWith(expect.anything(), {
-        projectId: 'project1',
-      });
-    });
-
-    it('should handle ZodError and return validation error message', async () => {
-      // Mock a function to throw a ZodError
-      const zodError = new Error('Expected number, received string');
-      zodError.name = 'ZodError';
-
-      // Mock getWorkItem to simulate throwing a validation error
-      (getWorkItem as jest.Mock).mockImplementationOnce(() => {
-        throw zodError;
-      });
-
-      const response = await callToolHandler({
-        params: {
-          name: 'get_work_item',
-          arguments: { workItemId: 'string-instead-of-number' },
-        },
-      });
-
-      expect(response.content[0].text).toContain(
-        'Expected number, received string',
-      );
-    });
-
-    it('should handle AzureDevOpsError and format the error message', async () => {
-      // Make listProjects throw an AzureDevOpsError
-      (listProjects as jest.Mock).mockImplementationOnce(() => {
-        throw new AzureDevOpsError('Test error');
-      });
-
-      const response = await callToolHandler({
-        params: {
-          name: 'list_projects',
-          arguments: { top: 10 },
-        },
-      });
-
-      expect(response.content[0].text).toContain(
-        'Azure DevOps API Error: Test error',
-      );
-    });
-
-    it('should handle AzureDevOpsValidationError and format the error message', async () => {
-      // Make listProjects throw an AzureDevOpsValidationError
-      (listProjects as jest.Mock).mockImplementationOnce(() => {
-        throw new AzureDevOpsValidationError('Validation failed');
-      });
-
-      const response = await callToolHandler({
-        params: {
-          name: 'list_projects',
-          arguments: { top: 10 },
-        },
-      });
-
-      expect(response.content[0].text).toContain(
-        'Validation Error: Validation failed',
-      );
-    });
-
-    it('should handle AzureDevOpsResourceNotFoundError and format the error message', async () => {
-      // Make listProjects throw an AzureDevOpsResourceNotFoundError
-      (listProjects as jest.Mock).mockImplementationOnce(() => {
-        throw new AzureDevOpsResourceNotFoundError('Resource not found');
-      });
-
-      const response = await callToolHandler({
-        params: {
-          name: 'list_projects',
-          arguments: { top: 10 },
-        },
-      });
-
-      expect(response.content[0].text).toContain(
-        'Not Found: Resource not found',
-      );
-    });
-
-    it('should handle AzureDevOpsAuthenticationError and format the error message', async () => {
-      // Make listProjects throw an AzureDevOpsAuthenticationError
-      (listProjects as jest.Mock).mockImplementationOnce(() => {
-        throw new AzureDevOpsAuthenticationError('Authentication failed');
-      });
-
-      const response = await callToolHandler({
-        params: {
-          name: 'list_projects',
-          arguments: { top: 10 },
-        },
-      });
-
-      expect(response.content[0].text).toContain(
-        'Authentication Failed: Authentication failed',
-      );
-    });
-
-    it('should handle generic errors and format the error message', async () => {
-      // Make listProjects throw a generic Error
-      (listProjects as jest.Mock).mockImplementationOnce(() => {
-        throw new Error('Generic error');
-      });
-
-      const response = await callToolHandler({
-        params: {
-          name: 'list_projects',
-          arguments: { top: 10 },
-        },
-      });
-
-      expect(response.content[0].text).toContain('Error: Generic error');
-    });
-
-    it('should handle create_work_item tool call', async () => {
-      // Mock the workitems.createWorkItem function
-      const mockWorkItem = {
-        id: 123,
-        fields: { 'System.Title': 'Test Work Item' },
-      };
-      (createWorkItem as jest.Mock).mockResolvedValueOnce(mockWorkItem);
-
-      // Call the handler with create_work_item parameters
-      const request = {
-        params: {
-          name: 'create_work_item',
-          arguments: {
-            projectId: 'testproject',
-            workItemType: 'Task',
-            title: 'Test Work Item',
-            description: 'This is a test work item',
-          },
-        },
-      };
-
-      const response = await callToolHandler(request as any);
-
-      // Verify the response
-      expect(response).toEqual({
-        content: [
-          { type: 'text', text: JSON.stringify(mockWorkItem, null, 2) },
-        ],
-      });
-
-      // Verify the createWorkItem function was called with the correct parameters
-      expect(createWorkItem).toHaveBeenCalledWith(
-        expect.anything(),
-        'testproject',
-        'Task',
-        {
-          title: 'Test Work Item',
-          description: 'This is a test work item',
-        },
-      );
-    });
-  });
-});
